Intial commit for netspace project
This is mostly a collection of scripts to set up a seperate network
namespace with some services running.

netspace-create is used to create a new namespace given an interface and
an address.

netspace-ctl is used for starting and stopping services inside the
created namespace. It  will run scripts in /root/.config/netspace/init.d
by default. Currently, stopping all services at once is supported.

netpspace-delete simply removes the namespace, and does not restore
initial state of the interface.

Example:
	sudo netspace-create eth2 198.18.11.99 netspace
	sudo netspace-ctl enter netspace

Signed-off-by: Patrik Lundgren <patrik.lundgren.95@gmail.com>